Phase 1 - add core module with no Spring deps
Factor out test utilities into reusable `TestFixtures`, streamline serde and processor tests

- Extract common test utilities (`driverConfig`, `TestRecord`) to new `TestFixtures` class for reusability.
- Use `TestFixtures.driverConfig` across specs to eliminate duplication.
- Simplify `BufferedRecordListSerde` initialization by replacing direct serde calls with dedicated serializer/deserializer fields.
- Minor refactor to improve readability in `ResequenceProcessor`.
